Move helper code to helper file
